KAN-22 패키지 구조 변경
