+++ src/main/java/main/sulsul/beverage/domain/RecordBeverage.java
@@ -1,4 +1,4 @@
-package main.sulsul.beverage;
+package main.sulsul.beverage.domain;
 
 import jakarta.persistence.*;
 import lombok.Getter;
