+++ src/main/java/main/sulsul/beverage/domain/dao/BeverageRecordRepository.java
@@ -1,6 +1,6 @@
-package main.sulsul.beverage.dao;
+package main.sulsul.beverage.domain.dao;
 
-import main.sulsul.beverage.RecordBeverage;
+import main.sulsul.beverage.domain.RecordBeverage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface BeverageRecordRepository extends JpaRepository<RecordBeverage, Long> {
