+++ src/main/java/main/sulsul/beverage/domain/dao/BeverageRepository.java
@@ -1,6 +1,6 @@
-package main.sulsul.beverage.dao;
+package main.sulsul.beverage.domain.dao;
 
-import main.sulsul.beverage.Beverage;
+import main.sulsul.beverage.domain.Beverage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface BeverageRepository extends JpaRepository<Beverage, Long> {
