+++ src/main/java/main/sulsul/beverage/domain/dao/RecordRepository.java
@@ -1,6 +1,6 @@
-package main.sulsul.beverage.dao;
+package main.sulsul.beverage.domain.dao;
 
-import main.sulsul.beverage.Record;
+import main.sulsul.beverage.domain.Record;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RecordRepository extends JpaRepository<Record, Long> {
