+++ src/main/java/main/sulsul/global/config/JpaAuditingConfiguration.java
@@ -1,0 +1,9 @@
+package main.sulsul.global.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaAuditingConfiguration {
+}
