--- src/main/java/main/sulsul/member/Member.java
@@ -1,21 +1,0 @@
-package main.sulsul.member;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.Getter;
-import main.sulsul.global.domain.BaseEntity;
-
-@Getter
-@Entity
-public class Member extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String username;
-
-    private String password;
-}
