+++ src/main/java/main/sulsul/member/domain/dao/MemberRepository.java
@@ -1,6 +1,6 @@
-package main.sulsul.member.dao;
+package main.sulsul.member.domain.dao;
 
-import main.sulsul.member.Member;
+import main.sulsul.member.domain.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
