+++ src/main/resources/db/migration/V20240114_2205__create_member_table.sql
@@ -6,5 +6,5 @@ CREATE TABLE member
     created_at DATETIME DEFAULT NOW(),
     updated_at DATETIME,
 
-    CONSTRAINT idx_username UNIQUE (username)
+    CONSTRAINT idx__username UNIQUE (username)
 );
