+++ src/test/java/main/sulsul/beverage/CategoryTest.java
@@ -1,8 +1,7 @@
 package main.sulsul.beverage;
 
+import main.sulsul.beverage.domain.Category;
 import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 class CategoryTest {
