+++ src/test/java/main/sulsul/beverage/dao/BeverageRepositoryTest.java
@@ -1,14 +1,13 @@
 package main.sulsul.beverage.dao;
 
-import main.sulsul.beverage.Beverage;
-import main.sulsul.beverage.Category;
+import main.sulsul.beverage.domain.Beverage;
+import main.sulsul.beverage.domain.Category;
+import main.sulsul.beverage.domain.dao.BeverageRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
 import java.util.Optional;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 @SpringBootTest
 class BeverageRepositoryTest {
